reorg jobs_metadata.yml in pip installed lib and renamed files to have cleaner order: pandas_sql, pandas_api, spark_sql, spark_api
