+++ yaetos/scripts/copy/jobs_metadata_external.yml
@@ -15,12 +15,22 @@ jobs:
     inputs:
       some_events: {'path':"{base_path}/raw/wiki/{latest}/", 'type':'csv', 'df_type':'pandas'}
       other_events: {'path':"{base_path}/raw/wiki/{latest}/", 'type':'csv', 'df_type':'pandas'}
-    output: {'path':'{base_path}/frontroom/wiki_ex1_sql/{now}/dataset.csv', 'type':'csv', 'df_type':'pandas'}
+    output: {'path':'{base_path}/frontroom/wiki_ex1_pandas_sql/{now}/dataset.csv', 'type':'csv', 'df_type':'pandas'}
     dependencies: [examples/ex0_extraction_job.py]
+    spark_boot: False
+
+  examples/ex1_pandas_api_job.py:
+    description: "job loading and processing data with pandas. No spark involved."
+    inputs:
+      some_events: {'path':"{base_path}/raw/wiki/{latest}/", 'type':'csv', 'df_type':'pandas'}
+      other_events: {'path':"{base_path}/raw/wiki/{latest}/", 'type':'csv', 'df_type':'pandas', 'read_kwargs':{}}
+    output: {'path':'{base_path}/frontroom/wiki_ex1_pandas_api/{now}/dataset.csv', 'type':'csv', 'df_type':'pandas', 'save_kwargs':{'sep':'|'}}
+    spark_boot: False
 
   examples/ex1_sql_spark_job:
     description: "shows a sql job, running on top of Spark."
     py_job: 'jobs/generic/sql_spark_job.py'
+    sql_file: 'jobs/examples/ex1_sql_job.sql'
     inputs:
       some_events: {'path':"{base_path}/raw/wiki/{latest}/", 'type':'csv'}
       other_events: {'path':"{base_path}/raw/wiki/{latest}/", 'type':'csv'}
@@ -28,24 +38,16 @@ jobs:
     dependencies: [examples/ex0_extraction_job.py]
     repartition: 1
 
-  examples/ex1_frameworked_job.py:
+  examples/ex1_spark_api_job.py:
     description: "shows frameworked pyspark ops, same as ex1_sql_job but gives access to spark ops to expand on sql."
     inputs:
       some_events: {'path':"{base_path}/raw/wiki/{latest}/", 'type':'csv'}
       other_events: {'path':"{base_path}/raw/wiki/{latest}/", 'type':'csv'}
-    output: {'path':'{base_path}/frontroom/wiki_ex1_frameworked/{now}/', 'type':'csv'}
+    output: {'path':'{base_path}/frontroom/wiki_ex1_spark_api/{now}/', 'type':'csv'}
     dependencies: [examples/ex0_extraction_job.py]
     frequency: 1 day
     start_date: "{today}T07:00:00"
     owners: ['some_email@address.com']
-
-  examples/ex1_pandas_job.py:
-    description: "job loading and processing data with pandas. No spark involved."
-    inputs:
-      some_events: {'path':"{base_path}/raw/wiki/{latest}/", 'type':'csv', 'df_type':'pandas'}
-      other_events: {'path':"{base_path}/raw/wiki/{latest}/", 'type':'csv', 'df_type':'pandas', 'read_kwargs':{}}
-    output: {'path':'{base_path}/frontroom/wiki_ex1_pandas/{now}/dataset.csv', 'type':'csv', 'df_type':'pandas', 'save_kwargs':{'sep':'|'}}
-    spark_boot: False
 
   # Your jobs should be added here.
